dcp-845 move terra config, storage client and transfer client
